requirements.txt
python-docx
farsi-tools
openpyxl
regex